1. Renamed Massage.java to Message.java. 2. I replaced ByteBuffer in Message.java to byte[]. 3. I removed the contructor and getters method in Message.java and replaced them with Lombok.

+++ src/test/java/com/steps/MyStepdefs.java
@@ -150,7 +150,7 @@ public class MyStepdefs {
         try {
             this.handShakeResponse = PeerCommunicator.sendMessage(
                     peer.getIpAddress(),
-                    peer.getTcpPort(),
+                    peer.getTcpPort().intValue(),
                     this.handShakeRequest);
             return true;
         } catch (IOException e) {
